Move FN44 tests to Field Notices
